+++ embedded-graphics/src/pixelcolor/mod.rs
@@ -88,11 +88,13 @@ mod conversion;
 mod gray_color;
 pub mod raw;
 mod rgb_color;
+mod web_colors;
 
 pub use binary_color::*;
 pub use gray_color::*;
 use raw::RawData;
 pub use rgb_color::*;
+pub use web_colors::WebColors;
 
 /// Pixel color trait.
 ///