+++ embedded-graphics/CHANGELOG.md
@@ -36,6 +36,7 @@ Embedded Graphics is a `no_std` library for adding graphics features to display 
 - [#429](https://github.com/jamwaffles/embedded-graphics/pull/429) Added `ToBytes` trait to convert colors into byte arrays.
 - [#431](https://github.com/jamwaffles/embedded-graphics/pull/431) Added `MockDisplay::affected_area` to get the area affected by previous drawing operations.
 - [#450](https://github.com/jamwaffles/embedded-graphics/pull/450) Added `ColorConverted` and `DrawTargetExt::color_converted` to support color conversion for draw targets.
+- [#438](https://github.com/jamwaffles/embedded-graphics/pull/438) Added majority CSS web colors as associated `const`s to the RGB color types.
 
 ### Changed